fix: code style changes
